fix route, validate dto request

+++ demo/src/main/java/com/neo4j_ecom/demo/model/dto/response/ReviewResponse.java
@@ -15,5 +15,5 @@ public class ReviewResponse {
 
     String id;
     String content;
-    double rating;
+    float rating;
 }
